7th commit

+++ controller/LikeController.js
@@ -1,41 +1,65 @@
 // 2024-04-15 감준서
+const ensureAuthorization = require('../auth');
+const jwt = require('jsonwebtoken');
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
 
+
 const addLike = (req,res) => {
 
-    const {id} = req.params;
-    const {user_id} = req.body;
+    const book_id = req.params.id;
 
-    let sql = 'INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)';
-    let values = [user_id, id];
-    conn.query(sql, values,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
+    let authorization = ensureAuthorization(req,res);
 
-            return res.status(StatusCodes.OK).json(results);
-    })
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요"
+        });
+    }else if(authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }else{
+        let sql = 'INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)';
+        let values = [authorization.id, book_id];
+        conn.query(sql, values,
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+
+                return res.status(StatusCodes.OK).json(results);
+        })
+    }
 };
 
-// SELECT count(*) FROM Bookshop.likes WHERE liked_book_id=1;
 const removeLike = (req,res) => {
-    const {id} = req.params;
-    const {user_id} = req.body;
+    const book_id = req.params.id;
 
-    let sql = ' DELETE FROM Bookshop.likes WHERE user_id = ? AND liked_book_id = ?';
-    let values = [user_id, id];
-    conn.query(sql, values,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
+    let authorization = ensureAuthorization(req,res);
 
-            return res.status(StatusCodes.OK).json(results);
-    })
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요"
+        });
+    }else if(authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }else{
+        let sql = ' DELETE FROM Bookshop.likes WHERE user_id = ? AND liked_book_id = ?';
+        let values = [authorization.id, book_id];
+        conn.query(sql, values,
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+
+                return res.status(StatusCodes.OK).json(results);
+        })
+    }
     
 };
